split core::ptr module into multiple files
